# Add AutoModule and AutoCriterion (build any nn.Module and nn.Criterio…