--- a/src/auto/init.lua
+++ b/src/auto/init.lua
@@ -1,0 +1,4 @@
+local auto = {
+}
+
+return auto