--- a/src/init.lua
+++ b/src/init.lua
@@ -8,6 +8,9 @@ autograd.LICENSE = 'MIT'
 -- Sub packages:
 autograd.functionalize = require 'autograd.nnwrapper'
 autograd.nn = require('autograd.nnwrapper')('nn')
+autograd.nn.AutoModule = require 'autograd.auto.AutoModule'
+autograd.nn.AutoCriterion = require 'autograd.auto.AutoCriterion'
+autograd.auto = require 'autograd.auto'
 autograd.model = require 'autograd.model'
 autograd.loss = require 'autograd.loss'
 autograd.util = require 'autograd.util'