This patch enables AutoModule definition from any forward function with the specified format, even with non-scalar outputs. This required to separate the grad(...) function in two parts. All tests pass, benchmarks are similar, so this shouldn't impact the current state. This should enable easy adoption for people who have nn/nngraph code embedded into another package and need the nn API. 

```
Benchmark:
-- Before
lstm (ag+nn) =>  nn:  0.94s, autograd:  1.87s, ratio:  1.99x
mlp (ag, batched) =>  nn:  0.28s, autograd:  0.70s, ratio:  2.53x
lstm (ag+nn, batched) =>  nn:  2.21s, autograd:  5.58s, ratio:  2.52x
logistic (ag) =>  nn:  0.96s, autograd:  2.76s, ratio:  2.88x
cnn (nnWrap, batched) =>  nn:  1.36s, autograd:  1.40s, ratio:  1.03x
mlp (ag) =>  nn:  0.39s, autograd:  1.09s, ratio:  2.79x
mlp (nn+ag) =>  nn:  0.38s, autograd:  0.52s, ratio:  1.37x
cnn (nn, batched) =>  nn:  1.36s, autograd:  1.60s, ratio:  1.17x
mlp (nn, batched) =>  nn:  0.27s, autograd:  0.44s, ratio:  1.60x
mlp (nn) =>  nn:  0.44s, autograd:  0.48s, ratio:  1.10x
logistic (nn) =>  nn:  0.79s, autograd:  1.34s, ratio:  1.70x
mlp (ag, fprop) =>  nn:  1.68s, autograd:  1.93s, ratio:  1.15x

-- New
lstm (ag+nn) =>  nn:  0.96s, autograd:  1.94s, ratio:  2.03x
mlp (ag, batched) =>  nn:  0.28s, autograd:  0.70s, ratio:  2.51x
lstm (ag+nn, batched) =>  nn:  2.21s, autograd:  5.08s, ratio:  2.30x
logistic (ag) =>  nn:  0.96s, autograd:  2.84s, ratio:  2.95x
cnn (nnWrap, batched) =>  nn:  1.37s, autograd:  1.38s, ratio:  1.01x
mlp (ag) =>  nn:  0.40s, autograd:  1.09s, ratio:  2.72x
mlp (nn+ag) =>  nn:  0.39s, autograd:  0.53s, ratio:  1.35x
cnn (nn, batched) =>  nn:  1.36s, autograd:  1.66s, ratio:  1.22x
mlp (nn, batched) =>  nn:  0.30s, autograd:  0.43s, ratio:  1.44x
mlp (nn) =>  nn:  0.43s, autograd:  0.50s, ratio:  1.16x
logistic (nn) =>  nn:  0.78s, autograd:  1.43s, ratio:  1.82x
mlp (ag, fprop) =>  nn:  1.67s, autograd:  1.95s, ratio:  1.16x
```