@soumith 2015-11-07
once this is merged, please dont forget to send in a PR to torch/rocks to update the rockspec there (for people who install it with "luarocks install autograd" ). @clementfarabet this is the nice part of using CMake based install, the build can be controlled from the repo itself, and adding files doesn't need updates to the torch/rocks rockspec.