@clementfarabet 2015-11-07
@soumith do you mean that if not using `cmake` a 2nd call to `luarocks install autograd` will not trigger a re-install? But if using `cmake` it will? :-)