@soumith 2015-11-07
@clementfarabet I mean that:
after this PR is merged,

luarocks install autograd

will actually give errors, because it will be installing the rockspec in torch/rocks: https://github.com/torch/rocks/blob/master/autograd-scm-1.rockspec which wont install the new auto modules (but they will be referred to in the lua scripts).

One will have to carefully make sure that making any changes to the rockspec in this repo will update the rockspec in torch/rocks. 

With the CMake install, you typically change the CMakeLists.txt file, not the rockspec.