@clementfarabet 2015-11-08
@soumith oh yes absolutely, when we add new modules, sure. Might switch to cmake just to get *.lua.