@andresy 2015-11-08
@clementfarabet (following @soumith 's remark): say you have a

```lua
local autograd = {}
```

which may be coming from

```
local autograd = require 'autograd.env'
```
(where env.lua simply returns a table -- this avoids loops requiring autograd).

so now you can do:

```lua
local MyClass, ParentClass = torch.class('autograd.MyClass', 'nn.ParentClass', autograd)
```

In short: torch.class() makes acts locally, and put a constructor into the table provided as extra argument.

The same is possible in C with the new function pointed out by Soumith.