@clementfarabet 2015-11-08
This is very neat! ok so let's change the names to `autograd.Module` and `autograd.Criterion`.