@clementfarabet 2015-11-08
So question: when you pass a table to torch.class, what's the use of the scope in the first argument?