@andresy 2015-11-08
You mean 'foo' in 'foo.Bar'? The class name is a uniq ID to identify the class when it is on the lua stack. So good practice to namespace it by the package name, to avoid clashes.