@clementfarabet 2015-11-08
Got it, but so `foo` could be anything and has nothing to do with the local table?