@andresy 2015-11-08
Yes. It is up to you. You can call your class without prefix too, like Bar__ZEbulLonX22L, but it will be harder to remember.