@clementfarabet 2015-11-08
Ok makes sense. Bar__ZEbulLonX22L is pretty neat :-).

On Sat, Nov 7, 2015 at 7:50 PM, Ronan Collobert <notifications@github.com>
wrote:

> Yes. It is up to you. You can call your class without prefix too, like
> Bar__ZEbulLonX22L, but it will be harder to remember.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/twitter/torch-autograd/pull/12#issuecomment-154768144>
> .
>