@soumith 2015-11-08
https://github.com/soumith/Bar__ZEbulLonX22L.torch