@andresy 2015-11-08
Man, one has to be careful with those namespaces :smile: