@nkoumchatzky 2015-11-08
Still not sure what to do about AutoModule being an nn class. Unless I push everything to https://github.com/soumith/Bar__ZEbulLonX22L.torch and problem is solved :-)