@soumith 2015-11-08
@nkoumchatzky you can simply change the lines to:
local auto = require 'autograd.auto'
local AutoCriterion, parent = torch.class('autograd.Criterion', 'nn.Criterion', auto)

and change src/auto/init.lua

to:

local auto = {}
return auto