@andresy 2015-11-08
note: could be in src/auto.lua too

> and change src/auto/init.lua
> 
> to:
> 
> local auto = {}
> return auto