@mdenil 2015-11-08
Hi, I'm really happy to see this feature.

Would it be possible to have AutoModule accept an arbitrary table of parameters like a regular autograd function instead of being restricted to just 'weight' and 'bias'?