@nkoumchatzky 2015-11-09
I amended this PR:
- No arbitrary params for now because self.weight/gradWeight and self.bias/gradBias are built-in and this would require a lot of function overriding and could not work as expected on the user side. Maybe in another PR.
- require 'autograd.nn.AutoModule' returns a function(moduleName) that lets the user create a new module that lives locally under autograd.auto.moduleName
This enables the creation of arbitrary modules that are distinguishable from one another.