@clementfarabet 2015-11-09
Ok – it just needs a rebase on master (because of a change in the rockspec).