@nkoumchatzky 2015-11-10
Added doc to README.